fix: fix tests after rewriting form component
