+++ package.json
@@ -44,6 +44,7 @@
     "lint-staged": "^13.2.0",
     "prettier": "^2.8.4",
     "react-hook-form": "^7.43.8",
+    "react-test-renderer": "^18.2.0",
     "typescript": "^4.9.3",
     "vite": "^4.2.0",
     "vite-plugin-eslint": "^1.8.1",
