+++ src/components/main/forms/FormPage.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import './FormPage.scss';
 import Title from '../../header/Title';
-import Form from './NewForm';
+import Form from './Form';
 
 export default class FormPage extends React.Component {
   render() {
