+++ src/components/main/forms/UI/Button.tsx
@@ -2,9 +2,5 @@ import React from 'react';
 import { TProps } from '../../../../types/props-types';
 
 export default function Button(props: TProps) {
-  return (
-    <button className="button button__submit" type="submit">
-      {props.children}
-    </button>
-  );
+  return <input className="button button__submit" type="submit" value={'submit'}></input>;
 }
