+++ src/main.tsx
@@ -4,9 +4,4 @@ import App from './App';
 import './main.scss';
 
 export const root: HTMLDivElement | null = document.querySelector('#root');
-if (root)
-  ReactDOM.createRoot(root).render(
-    // <React.StrictMode>
-    <App />
-    // </React.StrictMode>
-  );
+if (root) ReactDOM.createRoot(root).render(<App />);
